+++ tests/unit/no-top-level-variables.test.ts
@@ -57,7 +57,11 @@ const invalid: RuleTester.InvalidTestCase[] = [
     `,
     errors: [
       {
-        messageId: 'message'
+        messageId: 'message',
+        line: 1,
+        column: 5,
+        endLine: 1,
+        endColumn: 16
       }
     ]
   },
@@ -68,7 +72,11 @@ const invalid: RuleTester.InvalidTestCase[] = [
     `,
     errors: [
       {
-        messageId: 'message'
+        messageId: 'message',
+        line: 1,
+        column: 5,
+        endLine: 1,
+        endColumn: 16
       }
     ]
   },
@@ -85,7 +93,11 @@ const invalid: RuleTester.InvalidTestCase[] = [
     ],
     errors: [
       {
-        messageId: 'message'
+        messageId: 'message',
+        line: 1,
+        column: 5,
+        endLine: 1,
+        endColumn: 19
       }
     ]
   },
@@ -95,7 +107,11 @@ const invalid: RuleTester.InvalidTestCase[] = [
     `,
     errors: [
       {
-        messageId: 'message'
+        messageId: 'message',
+        line: 1,
+        column: 5,
+        endLine: 1,
+        endColumn: 8
       }
     ]
   }